# Switch to zstd Rust library (C binding of the official libzstd)