--- a/wild_lib/Cargo.toml
+++ b/wild_lib/Cargo.toml
@@ -32,8 +32,8 @@ sharded-vec-writer = "0.1.0"
 itertools = "0.13.0"
 bytesize = "1.3.0"
 flate2 = "1.0.33"
-ruzstd = "0.7.1"
 bumpalo-herd = "0.1.2"
+zstd = "0.13.2"
 
 [dev-dependencies]
 ar = "0.9.0"